# backend/postman/openapi.json
{"openapi":"3.1.0","info":{"title":"SportsTicketPlatform API","description":"SportsTicketPlatform API - Phase 3 (Raw SQL & Redis Cache)","version":"3.0.0"},"paths":{"/api/auth/otp":{"post":{"tags":["Authentication"],"summary":"Request Otp","operationId":"request_otp_api_auth_otp_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/OTPRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/OTPResponse"}}}},"500":{"description":"Internal Server Error"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/auth/signup":{"post":{"tags":["Authentication"],"summary":"Signup","operationId":"signup_api_auth_signup_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserSignup"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TokenResponse"}}}},"400":{"description":"Invalid/Expired OTP or User/Email already exists"},"500":{"description":"Database error"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/auth/login":{"post":{"tags":["Authentication"],"summary":"Login","description":"Logging in using the standard OAuth2 form.\nUse your phone number as the username on Swagger.","operationId":"login_api_auth_login_post","requestBody":{"content":{"application/x-www-form-urlencoded":{"schema":{"$ref":"#/components/schemas/Body_login_api_auth_login_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TokenResponse"}}}},"401":{"description":"Invalid phone number or password"},"500":{"description":"Database error"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/auth/me/test-auth":{"get":{"tags":["Authentication","Authentication"],"summary":"Test Authentication","description":"This endpoint is for testing the JWT security lock.","operationId":"test_authentication_api_auth_me_test_auth_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"401":{"description":"Not authenticated"}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/tickets/search":{"get":{"tags":["Tickets"],"summary":"Advanced Ticket Search with Smart Redis Caching & Multi-Filtering","operationId":"search_tickets_api_tickets_search_get","parameters":[{"name":"sport_type","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"Sport type: football, volleyball, basketball","title":"Sport Type"},"description":"Sport type: football, volleyball, basketball"},{"name":"venue","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"Name of the venue/stadium","title":"Venue"},"description":"Name of the venue/stadium"},{"name":"min_price","in":"query","required":false,"schema":{"anyOf":[{"type":"number"},{"type":"null"}],"description":"Minimum ticket price","title":"Min Price"},"description":"Minimum ticket price"},{"name":"max_price","in":"query","required":false,"schema":{"anyOf":[{"type":"number"},{"type":"null"}],"description":"Maximum ticket price","title":"Max Price"},"description":"Maximum ticket price"},{"name":"team_name","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"Search by team name (home or away)","title":"Team Name"},"description":"Search by team name (home or away)"},{"name":"ticket_tier","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"Ticket tier: VIP, Normal, Premium","title":"Ticket Tier"},"description":"Ticket tier: VIP, Normal, Premium"},{"name":"start_date","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"Matches starting from date (YYYY-MM-DD)","title":"Start Date"},"description":"Matches starting from date (YYYY-MM-DD)"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TicketListResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/tickets/{ticket_id}":{"get":{"tags":["Tickets"],"summary":"Get ticket details with JOINs and COALESCE","operationId":"get_ticket_details_api_tickets__ticket_id__get","parameters":[{"name":"ticket_id","in":"path","required":true,"schema":{"type":"integer","title":"Ticket Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TicketDetailResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/reservations/":{"post":{"tags":["Reservations"],"summary":"Reserve a ticket with 15-min lock and background reminder","operationId":"reserve_ticket_api_reservations__post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReservationRequest"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/payments/":{"post":{"tags":["Payments"],"summary":"Process payment for a pending reservation","operationId":"process_payment_api_payments__post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PaymentRequest"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PaymentResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/payments/cancellation-penalty/{reservation_id}":{"get":{"tags":["Payments"],"summary":"Calculate penalty for cancelling a ticket","operationId":"calculate_cancellation_penalty_api_payments_cancellation_penalty__reservation_id__get","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"reservation_id","in":"path","required":true,"schema":{"type":"integer","title":"Reservation Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/payments/cancel":{"post":{"tags":["Payments"],"summary":"Cancel a paid reservation and process refund","operationId":"cancel_ticket_api_payments_cancel_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CancelTicketRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/user/bookings":{"get":{"tags":["User Profile"],"summary":"Get booking history for the current user","operationId":"get_user_bookings_api_user_bookings_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/BookingResponse"},"type":"array","title":"Response Get User Bookings Api User Bookings Get"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/user/profile":{"put":{"tags":["User Profile"],"summary":"Update user profile and invalidate cache","operationId":"update_profile_api_user_profile_put","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserProfileUpdate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/reports/":{"get":{"tags":["Reports & Support"],"summary":"Get all reports submitted by the current user","operationId":"get_user_reports_api_reports__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/ReportResponse"},"type":"array","title":"Response Get User Reports Api Reports  Get"}}}}},"security":[{"OAuth2PasswordBearer":[]}]},"post":{"tags":["Reports & Support"],"summary":"Submit a new support ticket/report","operationId":"create_report_api_reports__post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReportCreate"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReportResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/admin/dashboard-stats":{"get":{"tags":["Admin Dashboard"],"summary":"Get aggregated statistics for the admin dashboard","operationId":"get_dashboard_statistics_api_admin_dashboard_stats_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DashboardStatsResponse"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/admin/tickets":{"get":{"tags":["Admin Dashboard"],"summary":"View all reservations and suspicious transactions (Support/Admin)","operationId":"get_all_reservations_api_admin_tickets_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/AdminReservationResponse"},"type":"array","title":"Response Get All Reservations Api Admin Tickets Get"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/admin/manage":{"put":{"tags":["Admin Dashboard"],"summary":"Change status of reports or cancel/approve reservations","operationId":"manage_entity_api_admin_manage_put","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AdminManageRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/cities-venues":{"get":{"tags":["Locations & Venues"],"summary":"Get unique list of cities and venues","operationId":"get_cities_and_venues_api_cities_venues_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/":{"get":{"tags":["System Health"],"summary":"Health Check","operationId":"health_check__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"AdminManageRequest":{"properties":{"entity_type":{"type":"string","title":"Entity Type"},"entity_id":{"type":"integer","title":"Entity Id"},"new_status":{"type":"string","title":"New Status"}},"type":"object","required":["entity_type","entity_id","new_status"],"title":"AdminManageRequest"},"AdminReservationResponse":{"properties":{"reservation_id":{"type":"integer","title":"Reservation Id"},"user_id":{"type":"integer","title":"User Id"},"first_name":{"type":"string","title":"First Name"},"last_name":{"type":"string","title":"Last Name"},"phone_number":{"type":"string","title":"Phone Number"},"ticket_id":{"type":"integer","title":"Ticket Id"},"venue_name":{"type":"string","title":"Venue Name"},"match_date":{"type":"string","format":"date-time","title":"Match Date"},"status":{"type":"string","title":"Status"},"payment_amount":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Payment Amount"}},"type":"object","required":["reservation_id","user_id","first_name","last_name","phone_number","ticket_id","venue_name","match_date","status"],"title":"AdminReservationResponse"},"Body_login_api_auth_login_post":{"properties":{"grant_type":{"anyOf":[{"type":"string","pattern":"password"},{"type":"null"}],"title":"Grant Type"},"username":{"type":"string","title":"Username"},"password":{"type":"string","title":"Password"},"scope":{"type":"string","title":"Scope","default":""},"client_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Client Id"},"client_secret":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Client Secret"}},"type":"object","required":["username","password"],"title":"Body_login_api_auth_login_post"},"BookingResponse":{"properties":{"reservation_id":{"type":"integer","title":"Reservation Id"},"ticket_id":{"type":"integer","title":"Ticket Id"},"home_team":{"type":"string","title":"Home Team"},"away_team":{"type":"string","title":"Away Team"},"match_date":{"type":"string","format":"date-time","title":"Match Date"},"reservation_status":{"type":"string","title":"Reservation Status"},"payment_status":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Payment Status"},"amount_paid":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Amount Paid"},"reserved_at":{"type":"string","format":"date-time","title":"Reserved At"}},"type":"object","required":["reservation_id","ticket_id","home_team","away_team","match_date","reservation_status","payment_status","amount_paid","reserved_at"],"title":"BookingResponse"},"CancelTicketRequest":{"properties":{"reservation_id":{"type":"integer","title":"Reservation Id"}},"type":"object","required":["reservation_id"],"title":"CancelTicketRequest"},"DashboardStatsResponse":{"properties":{"total_revenue":{"type":"number","title":"Total Revenue"},"total_tickets_sold":{"type":"integer","title":"Total Tickets Sold"},"total_cancellations":{"type":"integer","title":"Total Cancellations"},"pending_reports":{"type":"integer","title":"Pending Reports"}},"type":"object","required":["total_revenue","total_tickets_sold","total_cancellations","pending_reports"],"title":"DashboardStatsResponse"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"OTPRequest":{"properties":{"phone_number":{"type":"string","title":"Phone Number"}},"type":"object","required":["phone_number"],"title":"OTPRequest"},"OTPResponse":{"properties":{"message":{"type":"string","title":"Message"},"otp":{"type":"string","title":"Otp"},"expires_in":{"type":"string","title":"Expires In"}},"type":"object","required":["message","otp","expires_in"],"title":"OTPResponse"},"PaymentRequest":{"properties":{"reservation_id":{"type":"integer","title":"Reservation Id"},"payment_method":{"type":"string","title":"Payment Method","default":"credit_card"}},"type":"object","required":["reservation_id"],"title":"PaymentRequest"},"PaymentResponse":{"properties":{"payment_id":{"type":"integer","title":"Payment Id"},"reservation_id":{"type":"integer","title":"Reservation Id"},"amount":{"type":"number","title":"Amount"},"status":{"type":"string","title":"Status"},"message":{"type":"string","title":"Message"},"paid_at":{"type":"string","format":"date-time","title":"Paid At"}},"type":"object","required":["payment_id","reservation_id","amount","status","message","paid_at"],"title":"PaymentResponse"},"ReportCreate":{"properties":{"category":{"type":"string","title":"Category"},"report_text":{"type":"string","title":"Report Text"},"ticket_id":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Ticket Id"},"reservation_id":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Reservation Id"}},"type":"object","required":["category","report_text"],"title":"ReportCreate"},"ReportResponse":{"properties":{"report_id":{"type":"integer","title":"Report Id"},"user_id":{"type":"integer","title":"User Id"},"ticket_id":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Ticket Id"},"reservation_id":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Reservation Id"},"category":{"type":"string","title":"Category"},"report_text":{"type":"string","title":"Report Text"},"status":{"type":"string","title":"Status"},"created_at":{"type":"string","format":"date-time","title":"Created At"}},"type":"object","required":["report_id","user_id","category","report_text","status","created_at"],"title":"ReportResponse"},"ReservationRequest":{"properties":{"ticket_id":{"type":"integer","title":"Ticket Id"}},"type":"object","required":["ticket_id"],"title":"ReservationRequest"},"TicketDetailResponse":{"properties":{"ticket_id":{"type":"integer","title":"Ticket Id"},"title":{"type":"string","title":"Title"},"sport_type":{"type":"string","title":"Sport Type"},"home_team":{"type":"string","title":"Home Team"},"away_team":{"type":"string","title":"Away Team"},"venue_name":{"type":"string","title":"Venue Name"},"city":{"type":"string","title":"City"},"ticket_tier":{"type":"string","title":"Ticket Tier"},"organizer":{"type":"string","title":"Organizer"},"match_date":{"type":"string","format":"date-time","title":"Match Date"},"price":{"type":"number","title":"Price"},"remaining_capacity":{"type":"integer","title":"Remaining Capacity"},"is_active":{"type":"boolean","title":"Is Active"},"league_name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"League Name"},"facility_name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Facility Name"},"seat_section":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Seat Section"},"row_number":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Row Number"},"seat_number":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Seat Number"},"specific_ticket_tier":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Specific Ticket Tier"},"amenities":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Amenities"}},"type":"object","required":["ticket_id","title","sport_type","home_team","away_team","venue_name","city","ticket_tier","organizer","match_date","price","remaining_capacity","is_active"],"title":"TicketDetailResponse"},"TicketListResponse":{"properties":{"source":{"type":"string","title":"Source"},"count":{"type":"integer","title":"Count"},"tickets":{"items":{"$ref":"#/components/schemas/TicketResponse"},"type":"array","title":"Tickets"}},"type":"object","required":["source","count","tickets"],"title":"TicketListResponse"},"TicketResponse":{"properties":{"ticket_id":{"type":"integer","title":"Ticket Id"},"title":{"type":"string","title":"Title"},"sport_type":{"type":"string","title":"Sport Type"},"home_team":{"type":"string","title":"Home Team"},"away_team":{"type":"string","title":"Away Team"},"venue_name":{"type":"string","title":"Venue Name"},"city":{"type":"string","title":"City"},"ticket_tier":{"type":"string","title":"Ticket Tier"},"organizer":{"type":"string","title":"Organizer"},"match_date":{"type":"string","format":"date-time","title":"Match Date"},"price":{"type":"number","title":"Price"},"remaining_capacity":{"type":"integer","title":"Remaining Capacity"},"is_active":{"type":"boolean","title":"Is Active"}},"type":"object","required":["ticket_id","title","sport_type","home_team","away_team","venue_name","city","ticket_tier","organizer","match_date","price","remaining_capacity","is_active"],"title":"TicketResponse"},"TokenResponse":{"properties":{"access_token":{"type":"string","title":"Access Token"},"token_type":{"type":"string","title":"Token Type"},"message":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Message"}},"type":"object","required":["access_token","token_type"],"title":"TokenResponse"},"UserProfileUpdate":{"properties":{"first_name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"First Name"},"last_name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Last Name"},"city":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"City"}},"type":"object","title":"UserProfileUpdate"},"UserSignup":{"properties":{"phone_number":{"type":"string","title":"Phone Number"},"email":{"type":"string","format":"email","title":"Email"},"password":{"type":"string","title":"Password"},"otp_code":{"type":"string","title":"Otp Code"},"first_name":{"type":"string","title":"First Name"},"last_name":{"type":"string","title":"Last Name"},"city":{"type":"string","title":"City"}},"type":"object","required":["phone_number","email","password","otp_code","first_name","last_name","city"],"title":"UserSignup"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}},"securitySchemes":{"OAuth2PasswordBearer":{"type":"oauth2","flows":{"password":{"scopes":{},"tokenUrl":"/api/auth/login"}}}}}}